+++ example/lib/counter_simple_cube/counter_simple_cube.dart
@@ -1,7 +1,7 @@
 import 'package:cubes/cubes.dart';
 
 class CounterSimpleCube extends SimpleCube {
-  final count = 0.obsValue;
+  final count = 0.obs;
 
   void increment() {
     count.modify((value) => value + 1);